Updated main.c
Added UART code to print "Hello World" once a second.

Also added a a define around SystemClock_Config() as there is a bug
with RENODE and the STM32F446 that will make the RENODE never return
from SystemClock_Config(). To fix this, SystemClock_Config() contents
are wrapped a #if/#endif.